rename endpoint etc.
