+++ server/api/catalog-view.js
