+++ server/api/index.js
@@ -1,20 +1,22 @@
-const catalog = require('./catalog');
+const catalogView = require('./catalog-view');
 const upload = require('./upload');
 const catalogUserId = require('./catalog-user-id');
 const postsViewId = require('./posts-view-id');
-const postsSearch = require('./posts-search');
 const postsDownloadId = require('./posts-download-id');
 const postsEditId = require('./posts-edit-id');
 const authSignUp = require('./auth-sign-up');
 const authSignIn = require('./auth-sign-in');
+const catalogSearch = require('./catalog-search');
 
 module.exports = {
-  catalog,
-  catalogUserId,
+  catalog: {
+    view: catalogView,
+    userId: catalogUserId,
+    search: catalogSearch
+  },
   upload,
   posts: {
     viewId: postsViewId,
-    search: postsSearch,
     downloadId: postsDownloadId,
     editId: postsEditId
   },
